backend/requirements.txt
fastapi
uvicorn
paho-mqtt
aiokafka
kafka-python
websockets
requests>=2.28.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
python-dotenv>=1.0.0
fastapi
uvicorn
paho-mqtt
aiokafka
kafka-python
websockets
